requirements.txt
python-dotenv==1.0.1
python-telegram-bot==20.8
requests==2.25.1
google-cloud==0.34.0
fastapi==0.115.4
uvicorn==0.32.0